add intervals ;move tests to test/ dir; move test classes to separate files
