+++ fava_investor/modules/performance/test/test_balances.py
@@ -2,7 +2,7 @@ from beancount.utils import test_utils
 from freezegun import freeze_time
 
 from fava_investor.modules.performance.balances import get_balances_tree
-from fava_investor.modules.performance.test_split import SplitTestCase, get_ledger
+from fava_investor.modules.performance.test.testutils import SplitTestCase, get_ledger
 
 CONFIG = {"accounts_pattern": "^Assets:Account"}
